requirements.txt
Django==2.0.7
django-webpack-loader==0.2.4
djangorestframework==3.8.2
wheel==0.26.0
gunicorn==19.5.0
whitenoise==2.0.6
requests
django-cors-headers
cassandra-driver==3.13.0
django-cassandra-engine
spotipy==2.4.4
scikit-learn
matplotlib
numpy==1.15.0
scipy
bs4
